+++ ext_tables.sql
@@ -3,7 +3,8 @@
 #
 CREATE TABLE fe_users (
 	tx_seminars_registration int(11) unsigned DEFAULT '0' NOT NULL,
-	default_organizer        int(11) unsigned DEFAULT '0' NOT NULL
+	default_organizer        int(11) unsigned DEFAULT '0' NOT NULL,
+	available_topics         text
 );